refactor: renamed to setup, added warning and tooltip
