+++ Editor/Setup/RigLoaderSceneSetup.cs.meta
