+++ Runtime/Setup/InteractionRigSetup.cs.meta
